metrics: add metrics definitions.
