+++ pkg/cke/main.go
@@ -24,7 +24,7 @@ var (
 	flgInterval        = pflag.String("interval", "1m", "check interval")
 	flgCertsGCInterval = pflag.String("certs-gc-interval", "1h", "tidy interval for expired certificates")
 	flgSessionTTL      = pflag.String("session-ttl", "60s", "leader session's TTL")
-	flgMetrics         = pflag.String("metrics", "0.0.0.0:10081", "<Listen IP>:<Port number>")
+	flgMetrics         = pflag.String("metrics", "0.0.0.0:10181", "<Listen IP>:<Port number>")
 	flgMetricsInterval = pflag.String("metrics-interval", "30s", "interval duration to collect metrics data")
 	flgDebugSabakan    = pflag.Bool("debug-sabakan", false, "debug sabakan integration")
 )
@@ -69,6 +69,11 @@ func main() {
 		log.ErrorExit(err)
 	}
 
+	metricsInterval, err := time.ParseDuration(*flgMetricsInterval)
+	if err != nil {
+		log.ErrorExit(err)
+	}
+
 	ttl, err := time.ParseDuration(*flgSessionTTL)
 	if err != nil {
 		log.ErrorExit(err)
@@ -107,12 +112,7 @@ func main() {
 	well.Go(controller.Run)
 
 	// Metrics
-	minterval, err := time.ParseDuration(*flgMetricsInterval)
-	if err != nil {
-		log.ErrorExit(err)
-	}
-
-	metricsUpdater := metrics.NewUpdater(minterval, etcd)
+	metricsUpdater := metrics.NewUpdater(metricsInterval, etcd)
 	well.Go(metricsUpdater.UpdateLoop)
 
 	metricsHandler := metrics.GetHandler()
